TrainingCategoryLabelPresenter の移動もれ対応
