+++ treco-web/src/app/(header)/(auth)/home/categories/[categoryId]/events/_components/training-category-label.tsx
@@ -1,7 +1,6 @@
-import React from 'react';
+import { TrainingMark } from '@/components/training-mark';
 
 import { cachedQueryCategory } from '../_queries/queries';
-import { TrainingCategoryLabelPresenter } from './training-events';
 
 type TrainingCategoryLabelContainerProps = {
   categoryId: string;
@@ -16,5 +15,20 @@ export async function TrainingCategoryLabelContainer({
       color={category.color}
       name={category.name}
     />
+  );
+}
+
+type TrainingCategoryLabelPresenterProps = {
+  color: string;
+  name: string;
+};
+export function TrainingCategoryLabelPresenter({
+  color,
+  name,
+}: TrainingCategoryLabelPresenterProps) {
+  return (
+    <div className="mb-4 flex items-center gap-2" data-testid="category-name">
+      <TrainingMark color={color} size="small" /> {name}
+    </div>
   );
 }
