+++ treco-web/src/app/(header)/(auth)/home/categories/[categoryId]/events/_components/training-events.tsx
@@ -1,25 +1,9 @@
-import { TrainingMark } from '@/components/training-mark';
 import { getSignedInTraineeId } from '@/lib/trainee';
 
 import { createNewRecordAction } from '../_actions';
 import { cachedQueryCategory, queryTrainingEvents } from '../_queries/queries';
 import { EventDelete } from './event-delete';
 import { EventEdit } from './event-edit';
-
-type TrainingCategoryLabelPresenterProps = {
-  color: string;
-  name: string;
-};
-export function TrainingCategoryLabelPresenter({
-  color,
-  name,
-}: TrainingCategoryLabelPresenterProps) {
-  return (
-    <div className="mb-4 flex items-center gap-2" data-testid="category-name">
-      <TrainingMark color={color} size="small" /> {name}
-    </div>
-  );
-}
 
 type TrainingEventsContainerProps = {
   categoryId: string;
